+++ packages/figma/src/codegen/targets/react/pipeline.ts
@@ -14,6 +14,7 @@ import {
   createIconSelfLayoutPropsConverter,
   createRadiusPropsConverter,
   createSelfLayoutPropsConverter,
+  createShadowPropsConverter,
   createShapeFillPropsConverter,
   createStrokePropsConverter,
   createTextFillPropsConverter,
@@ -27,9 +28,10 @@ import {
 } from "./shape";
 import { createTextTransformer } from "./text";
 import {
+  defaultEffectStyleNameFormatter,
+  defaultFillStyleResolver,
   defaultRawValueFormatters,
   defaultTextStyleNameFormatter,
-  defaultFillStyleResolver,
   defaultVariableNameFormatter,
 } from "./value-resolver";
 
@@ -53,6 +55,7 @@ export function createPipeline(options: CreatePipelineConfig = {}) {
     variableNameFormatter: defaultVariableNameFormatter,
     styleService,
     textStyleNameFormatter: defaultTextStyleNameFormatter,
+    effectStyleNameFormatter: defaultEffectStyleNameFormatter,
     fillStyleResolver: defaultFillStyleResolver,
     rawValueFormatters: defaultRawValueFormatters,
     shouldInferVariableName,
@@ -67,6 +70,7 @@ export function createPipeline(options: CreatePipelineConfig = {}) {
   const vectorChildrenFillPropsConverter = createVectorChildrenFillPropsConverter(valueResolver);
   const radiusPropsConverter = createRadiusPropsConverter(valueResolver);
   const strokePropsConverter = createStrokePropsConverter(valueResolver);
+  const shadowPropsConverter = createShadowPropsConverter(valueResolver);
   const typeStylePropsConverter = createTypeStylePropsConverter({
     valueResolver,
   });
@@ -80,6 +84,7 @@ export function createPipeline(options: CreatePipelineConfig = {}) {
     vectorChildrenFill: vectorChildrenFillPropsConverter,
     radius: radiusPropsConverter,
     stroke: strokePropsConverter,
+    shadow: shadowPropsConverter,
     typeStyle: typeStylePropsConverter,
   };