+++ packages/tailwind4-theme/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @seed-design/tailwind4-theme
 
+## 1.1.13
+
+### Patch Changes
+
+- cc4a45a: 신규 [Elevation 가이드](https://seed-design.io/docs/foundation/elevation)에 맞는 shadow 토큰을 추가합니다.
+
+  - React: Box, Flex, HStack 등 StyleProps를 사용하는 컴포넌트에서 `boxShadow` prop을 사용하여 shadow 토큰을 쉽게 사용할 수 있습니다.
+
 ## 1.1.6
 
 ### Patch Changes